refactor: Utils 패키지 생성 & 클래스 이동
