+++ README.md
@@ -145,6 +145,8 @@
 - [x] 메소드 길이 10 이하로 만들기
 - [x] 클래스 변수 2개 이하로 만들기
 - [x] 자판기 클래스 생성
+- [x] Utils 패키지 생성 및 클래스 이동
+- [ ] 컨버터 클래스에서 인풋 스트링 처리
 - [ ] MVC패턴 적용해 분할하기
 - [ ] 메소드가 한가지 일만 하도록 하기
 - [ ] InputController 속 메소드 반복/반복문 줄이고 구조화 하기
