+++ src/main/java/vendingmachine/Utils/Util.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.Utils;
 
 import java.util.Arrays;
