+++ src/main/java/vendingmachine/Utils/Validator/ProductValidator.java
@@ -1,11 +1,11 @@
-package vendingmachine.Model.Validator;
+package vendingmachine.Utils.Validator;
 
 import java.util.Arrays;
 
-import vendingmachine.Constants;
+import vendingmachine.Utils.Constants;
 
 public class ProductValidator {
-	public final String PRODUCT_STRING;
+	private final String PRODUCT_STRING;
 
 	public ProductValidator(String product) {
 		PRODUCT_STRING = product;
@@ -18,15 +18,15 @@ public class ProductValidator {
 	}
 
 	private void isRightString() {
-		if (!Constants.PRODUCT_PATTERN.matcher(PRODUCT_STRING).matches()) {
+		if (!Constants.PRODUCTS_PATTERN.matcher(PRODUCT_STRING).matches()) {
 			throw new IllegalArgumentException(Constants.ERROR_PRODUCT_STRING);
 		}
 	}
 
 	private void isNoDuplicate() {
 		String[] productNames = Arrays
-			.stream(PRODUCT_STRING.split(Constants.DELIMITER_PRODUCTS))
-			.map(product -> product.split(Constants.DELIMITER_PRODUCT)[0])
+			.stream(PRODUCT_STRING.split(Constants.SEPARATOR))
+			.map(product -> product.split(Constants.DELIMITER)[0])
 			.toArray(String[]::new);
 
 		if (productNames.length != Arrays.stream(productNames).distinct().count()) {
